+++ features/steps/gradle_plugin_steps.rb
@@ -44,25 +44,69 @@ steps %Q{
 }
 end
 
-Then(/^the request (\d+) is valid for the Android NDK Mapping API$/) do |request_index|
-  parts = find_request(request_index)[:body]
-  assert_not_nil(parts["soSymbolFile"], "'soSymbolFile' should not be nil")
-  assert_not_nil(parts["apiKey"], "'apiKey' should not be nil")
-  assert_not_nil(parts["sharedObjectName"], "'sharedObjectName' should not be nil")
-  assert_not_nil(parts["appId"], "'appId' should not be nil")
-  assert_not_nil(parts["arch"], "'arch' should not be nil")
+When("I set the fixture JVM arguments to {string}") do |jvm_args|
+steps %Q{
+  When I set environment variable "CUSTOM_JVM_ARGS" to "#{jvm_args}"
+}
 end
 
 When("I build the failing {string} using the {string} bugsnag config") do |module_config, bugsnag_config|
-  begin
-    steps %Q{
-      When I build "#{module_config}" using the "#{bugsnag_config}" bugsnag config
-    }
-    assert(false, "Expected script to fail with non-zero exit code")
-  rescue SystemExit
-  end
+  Runner.environment["MODULE_CONFIG"] = module_config
+  Runner.environment["BUGSNAG_CONFIG"] = bugsnag_config
+  _, exit_code = Runner.run_script("features/scripts/bundle_project_module.sh", blocking: true)
+  assert(exit_code != 0, "Expected script to fail with non-zero exit code, got #{exit_code}")
 end
 
 Then(/^the exit code equals (\d+)$/) do |exit_code|
   assert_equal(exit_code, $?.exitstatus.to_i)
+end
+
+Then('{int} requests are valid for the build API and match the following:') do |request_count, data_table|
+  requests = get_build_requests
+  assert_equal(request_count, requests.length, 'Wrong number of build API requests')
+  RequestSetAssertions.assert_requests_match requests, data_table
+end
+
+Then('{int} requests are valid for the android mapping API and match the following:') do |request_count, data_table|
+  requests = get_android_mapping_requests
+  assert_equal(request_count, requests.length, 'Wrong number of mapping API requests')
+  RequestSetAssertions.assert_requests_match requests, data_table
+end
+
+Then('{int} requests are valid for the android NDK mapping API and match the following:') do |request_count, data_table|
+  requests = get_android_ndk_mapping_requests
+  assert_equal(request_count, requests.length, 'Wrong number of NDK mapping API requests')
+  RequestSetAssertions.assert_requests_match requests, data_table
+end
+
+def valid_build_api?(request_body)
+  assert_equal($api_key, read_key_path(request_body, 'apiKey'))
+  assert_not_nil(read_key_path(request_body, 'appVersion'))
+  assert_not_nil(read_key_path(request_body, 'builderName'))
+  assert_not_nil(read_key_path(request_body, 'sourceControl.revision'))
+  assert_not_nil(read_key_path(request_body, 'metadata.os_name'))
+  assert_not_nil(read_key_path(request_body, 'metadata.os_arch'))
+  assert_not_nil(read_key_path(request_body, 'metadata.os_version'))
+  assert_not_nil(read_key_path(request_body, 'metadata.java_version'))
+  assert_not_nil(read_key_path(request_body, 'metadata.gradle_version'))
+  assert_not_nil(read_key_path(request_body, 'metadata.git_version'))
+end
+
+def valid_android_mapping_api?(request_body)
+  assert_equal($api_key, request_body['apiKey'])
+  assert_not_nil(request_body['proguard'])
+  assert_not_nil(request_body['appId'])
+  assert_not_nil(request_body['versionCode'])
+  assert_not_nil(request_body['buildUUID'])
+  assert_not_nil(request_body['versionName'])
+end
+
+# TODO can avoid duplication
+def valid_android_ndk_mapping_api?(request_body)
+  assert_equal($api_key, request_body['apiKey'])
+  assert_not_nil(request_body['soSymbolFile'])
+  assert_not_nil(request_body['appId'])
+  assert_not_nil(request_body['versionCode'])
+  assert_not_nil(request_body['buildUUID'])
+  assert_not_nil(request_body['versionName'])
 end