+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+## 5.2.0 (2020-09-22)
+
+* Support mapping upload from Android Gradle Plugin through a proxy
+  [#298](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/298)
+
+## 5.1.0 (2020-09-17)
+
+* Remove uploadDebugBuildMappings flag in favour of variantFilter
+  [#301](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/301)
+
+* Fix task ordering for upload task on AGP 3.5
+  [#300](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/300)
+
+* Fix bugsnag.sharedObjectPaths not uploading SO files
+  [#295](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/295)
+
+## 5.0.2 (2020-09-10)
+
+* Reword log messages that implied duplicate uploads
+  [#296](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/296)
+
+* Only set mappingFilesProvider on release task if obfuscation enabled
+  [#292](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/292)
+
+* Fix task dependencies and order for BugsnagManifestUuidTask in AGP <4.1.0
+  [#294](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/294)
+
+## 5.0.1 (2020-08-26)
+
+* Retry request by constructing new OkHttp request
+  [#278](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/278)
+
+* Apply `requestTimeoutMs` as an idle timeout
+  [#282](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/282)
+
+* Fix Bintray/jcenter publishing
+  [#283](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/283)
+
 ## 5.0.0 (2020-08-20)
 
 This release contains **breaking changes**. It contains numerous performance